requirements.txt
jsonschema>=4.22,<5.0
PyYAML>=6.0,<7.0
